+++ mobile/android/geckoview/src/androidTest/assets/web_extensions/test-support/background.js
@@ -59,6 +59,9 @@ const APIS = {
   FlushApzRepaints({ tab }) {
     return browser.test.flushApzRepaints(tab.id);
   },
+  ZoomToFocusedInput({ tab }) {
+    return browser.test.zoomToFocusedInput(tab.id);
+  },
   PromiseAllPaintsDone({ tab }) {
     return browser.test.promiseAllPaintsDone(tab.id);
   },